ReaderSpec renamed to DigitReaderSpec
