+++ Detectors/CTP/workflowIO/include/CTPWorkflowIO/DigitReaderSpec.h
